Move coniks hasher out
